[FIX] move settings to main package in molgenis-security and fix the molgenisinterceptor
