+++ molgenis-core-ui/src/main/java/org/molgenis/ui/MolgenisInterceptor.java
@@ -2,7 +2,7 @@ package org.molgenis.ui;
 
 import org.molgenis.data.i18n.LanguageService;
 import org.molgenis.data.settings.AppSettings;
-import org.molgenis.security.twofactor.settings.AuthenticationSettings;
+import org.molgenis.security.settings.AuthenticationSettings;
 import org.molgenis.ui.style.ThemeFingerprintRegistry;
 import org.molgenis.util.ResourceFingerprintRegistry;
 import org.molgenis.util.TemplateResourceUtils;
@@ -26,8 +26,8 @@ public class MolgenisInterceptor extends HandlerInterceptorAdapter
 	private final ResourceFingerprintRegistry resourceFingerprintRegistry;
 	private final ThemeFingerprintRegistry themeFingerprintRegistry;
 	private final TemplateResourceUtils templateResourceUtils;
-	private final AppSettings appSettings;
 	private final AuthenticationSettings authenticationSettings;
+	private final AppSettings appSettings;
 	private final String environment;
 	private final LanguageService languageService;
 
@@ -55,9 +55,8 @@ public class MolgenisInterceptor extends HandlerInterceptorAdapter
 			modelAndView.addObject(KEY_THEME_FINGERPRINT_REGISTRY, themeFingerprintRegistry);
 			modelAndView.addObject(KEY_RESOURCE_UTILS, templateResourceUtils);
 			modelAndView.addObject(KEY_APP_SETTINGS, appSettings);
+			modelAndView.addObject(KEY_AUTHENTICATION_SETTINGS, authenticationSettings);
 			modelAndView.addObject(KEY_ENVIRONMENT, getEnvironmentAttributes());
-			modelAndView.addObject(KEY_AUTHENTICATION_SETTINGS, authenticationSettings);
-			modelAndView.addObject(KEY_AUTHENTICATION_SETTINGS, authenticationSettings);
 			modelAndView.addObject(KEY_I18N, languageService.getBundle());
 		}
 	}
