+++ molgenis-security/src/main/java/org/molgenis/security/settings/AuthenticationSettingsImpl.java
@@ -1,4 +1,4 @@
-package org.molgenis.security.twofactor.settings;
+package org.molgenis.security.settings;
 
 import org.molgenis.data.settings.DefaultSettingsEntity;
 import org.molgenis.data.settings.DefaultSettingsEntityType;
