+++ molgenis-security/src/main/java/org/molgenis/security/twofactor/auth/TwoFactorAuthenticationFilter.java
@@ -4,7 +4,7 @@ import org.molgenis.security.core.utils.SecurityUtils;
 import org.molgenis.security.token.RestAuthenticationToken;
 import org.molgenis.security.twofactor.TwoFactorAuthenticationController;
 import org.molgenis.security.twofactor.service.TwoFactorAuthenticationService;
-import org.molgenis.security.twofactor.settings.AuthenticationSettings;
+import org.molgenis.security.settings.AuthenticationSettings;
 import org.molgenis.security.user.UserAccountService;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
